requirements.txt
requests
beautifulsoup4
lxml
python-docx
docx2pdf
tqdm
